requirements.txt
Keras==2.3.1
tensorflow-gpu==1.15.2
scikit-learn==0.23.1
Pillow
matplotlib
imutils
opencv-python
filterpy
